requirements.txt
pytorch
torchvision
matplotlib
ipython
jupyter
tqdm
click
numpy
tensorboard
pandas
